Debug via flags and campture image during run

+++ data_collector/cam.py
@@ -21,7 +21,7 @@ np.set_printoptions(threshold=sys.maxsize)
 motion_detected     = False
 keep_running        = True
 debug_file          = None
-debug_flag          = False
+flags               = None
 #Handle the SIGINT interrupt
 def signal_handler(signum, frame):
     global keep_running
@@ -46,16 +46,11 @@ class MotionDetec(array.PiMotionAnalysis):
                                                 ((0,0),(0,1)),
                                                 mode="constant",
                                                 constant_values=0)
-        #nx,ny                       = self.motion_mask.shape
-        #for i in range(nx):
-        #    for j in range(ny):
-        #        if j == ny-1:
-        #            print("{:.01f}".format(self.motion_mask[i,j]))
-        #        else:
-        #            print("{:.01f}".format(self.motion_mask[i,j]), end='')
 
     def analyse(self, a):
         global motion_detected
+        global flags
+        global debug_file
         a = np.sqrt(np.square(a['x'].astype(float)) +
                     np.square(a['y'].astype(float)))
 
@@ -66,8 +61,11 @@ class MotionDetec(array.PiMotionAnalysis):
                 (a > self.threshold).sum() < self.motion_mask.sum()*0.8:
             motion_detected         = True
             self.no_motion_frames   = 0
-            if debug_flag:
+            if flags[1]:
                 nx,ny                       = a.shape
+                debug_file.write("{}".format(dt.datetime.strftime(
+                                                    dt.now(),
+                                                    "%Y%m%d%H%M%S")))
                 for i in range(nx):
                     for j in range(ny):
                         if j == ny-1:
@@ -80,15 +78,15 @@ class MotionDetec(array.PiMotionAnalysis):
                 (a > self.threshold).sum() > self.num_blocks and \
                 (a > self.threshold).sum() < self.motion_mask.sum()*0.8:
             self.no_motion_frames   = 0
-            #if debug_flag:
-            #    nx,ny                       = a.shape
-            #    for i in range(nx):
-            #        for j in range(ny):
-            #            if j == ny-1:
-            #                debug_file.write("{:4.01f}\n".format(a[i,j]))
-            #            else:
-            #                debug_file.write("{:4.01f}".format(a[i,j]))
-            #    debug_file.write("\n")
+            if flags[1]:
+                nx,ny                       = a.shape
+                for i in range(nx):
+                    for j in range(ny):
+                        if j == ny-1:
+                            debug_file.write("{:4.01f}\n".format(a[i,j]))
+                        else:
+                            debug_file.write("{:4.01f}".format(a[i,j]))
+                debug_file.write("\n")
 
         elif    motion_detected         and \
                 (a > self.threshold).sum() <= self.num_blocks     and \
@@ -136,6 +134,15 @@ def read_status_file(name="/tmp/cam_flags"):
     flags   = flags == 1
     return flags
 
+def write_status_file(flgs,name="/tmp/cam_flags"):
+    out         = flgs.astype("int")
+    outstr      = ""
+    for fl in out[:-1]:
+        outstr += "{:d}\t".format(fl)
+    outstr     += "{:d}".format(out[-1])
+    with open(name,"w") as dafi:
+        dafi.write(outstr)
+
 def create_mask(camera,loglevel=1,praefix=""):
     if loglevel == 0:
         print("Some image will be taken")
@@ -160,9 +167,10 @@ def loop(   camera,
             buffer_time=120,
             motion_mask=np.ones((30,40)),
             motion_mask_st=np.ones((30,40))):
-
     global motion_detected
     global keep_running
+    global flags
+    global debug_file
 
     #setting up GPS buffer
     l_data_ti       = 0.
@@ -189,10 +197,18 @@ def loop(   camera,
                             motion_output=mclass)
     
     #Do some stuff while motion is not detected and wait
-    #start   = dt.now()
-    #while dt.now()-start < tidt(seconds=30.):
     while keep_running:
         flags       = read_status_file()
+        if flags[1] and debug_file == None:
+            debug_file      = open("/tmp/debug_cam_motion.txt","w")
+        elif not(flags[1]) and debug_file != None:
+            debug_file.close()
+        if flags[2]:
+            fname   = "{}{}".format(praefix,dt.datetime.strftime(dt.datetime.now(),"%Y%m%d_%H%M%S"))
+            camera.capture("{}.jpg",format(fname), use_video_port=True)
+            flags[2] = not(flags[2]
+            write_status_file(flags)
+
         if loglevel == 0:
             print("Waiting for motion")
             print("thresh={}, num_blocks={}".format(mclass.threshold,
@@ -201,8 +217,6 @@ def loop(   camera,
         camera.wait_recording(0.2)
         if motion_detected or flags[0]:
             fname   = "{}{}".format(praefix,dt.datetime.strftime(dt.datetime.now(),"%Y%m%d_%H%M%S"))
-            if debug_flag:
-                debug_file.write(fname+"\n")
             if loglevel < 2:
                 print("Motion at: {}".format(fname.split("/")[-1]))
             camera.split_recording("{}_during.mp4".format(fname),splitter_port=1)
@@ -287,13 +301,12 @@ if __name__ == "__main__":
     parser.add_option(  "", "--create_mask", dest="create_mask",
                         action="store_true",default=False,
                         help="Take an image for the creation of motion mask")
-    parser.add_option(  "", "--DEBUG", dest="DEBUG",
-                        action="store_true",default=False,
-                        help="Output motion array to file")
 
     (options, args) = parser.parse_args()
-    with open("/tmp/cam_flags","w") as fi:
-        fi.write("0\t0\t0")
+    if not(os.path.isfile("/tmp/cam_flags")):
+        with open("/tmp/cam_flags","w") as fi:
+            fi.write("0\t0\t0")
+    flags   = read_status_file()
 
     fname               = dt.datetime.strftime( dt.datetime.now(),"%Y%m%d")
     fname               = fname + options.postfix
@@ -307,12 +320,10 @@ if __name__ == "__main__":
 
     cam                 = init_camera(loglevel=int(options.loglevel))
 
-    debug_flag          = options.DEBUG
-    if debug_flag:
+    if flags[1]:
         debug_file      = open("/tmp/debug_cam_motion.txt","w")
         if loglevel == 0:
             print("Debug mode active")
-        
 
     if options.create_mask:
         create_mask(    cam,
@@ -353,5 +364,5 @@ if __name__ == "__main__":
                 motion_mask_st=mask_st)
 
     dinit_camera(cam)
-    if debug_flag:
+    if debug_file != None:
         debug_file.close()
